src/api/pyproject.toml
[tool.poetry]
name = "nekosapi"
version = "3.0.0"
description = "An open-source anime artwork REST API."
authors = ["Nyeki <hello@nyeki.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
django-ninja = "^0.22.2"
psycopg = {extras = ["binary"], version = "^3.1.12"}
python-dotenv = "^1.0.0"
orjson = "^3.9.9"
pillow = "^10.1.0"
uvicorn = "^0.23.2"
django-dynamic-filenames = "^1.3.2"
django-bunny = "^1.1.6"
django-cleanup = "^8.0.0"
colorthief = "^0.2.1"
imagehash = "^4.3.1"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
